+++ lib/waterdrop/clients/dummy.rb
@@ -5,44 +5,58 @@ module WaterDrop
     # A dummy client that is supposed to be used instead of Rdkafka::Producer in case we don't
     # want to dispatch anything to Kafka.
     #
-    # It does not store anything and just ignores messages.
+    # It does not store anything and just ignores messages. It does however return proper delivery
+    # handle that can be materialized into a report.
     class Dummy
+      # `::Rdkafka::Producer::DeliveryHandle` object API compatible dummy object
+      class Handle < ::Rdkafka::Producer::DeliveryHandle
+        # @param topic [String] topic where we want to dispatch message
+        # @param partition [Integer] target partition
+        # @param offset [Integer] offset assigned by our fake "Kafka"
+        def initialize(topic, partition, offset)
+          @topic = topic
+          @partition = partition
+          @offset = offset
+        end
+
+        # Does not wait, just creates the result
+        #
+        # @param _args [Array] anything the wait handle would accept
+        # @return [::Rdkafka::Producer::DeliveryReport]
+        def wait(*_args)
+          create_result
+        end
+
+        # Creates a delivery report with details where the message went
+        #
+        # @return [::Rdkafka::Producer::DeliveryReport]
+        def create_result
+          ::Rdkafka::Producer::DeliveryReport.new(
+            @partition,
+            @offset,
+            @topic
+          )
+        end
+      end
+
       # @param _producer [WaterDrop::Producer]
       # @return [Dummy] dummy instance
       def initialize(_producer)
-        @counter = -1
+        @counters = Hash.new { |h, k| h[k] = -1 }
       end
 
-      # Dummy method for returning the delivery report
-      # @param _args [Object] anything that the delivery handle accepts
-      # @return [::Rdkafka::Producer::DeliveryReport]
-      def wait(*_args)
-        ::Rdkafka::Producer::DeliveryReport.new(0, @counter += 1)
+      # "Produces" the message
+      # @param topic [String, Symbol] topic where we want to dispatch message
+      # @param partition [Integer] target partition
+      # @param _args [Hash] remaining details that are ignored in the dummy mode
+      # @return [Handle] delivery handle
+      def produce(topic:, partition: 0, **_args)
+        Handle.new(topic.to_s, partition, @counters["#{topic}#{partition}"] += 1)
       end
 
       # @param _args [Object] anything really, this dummy is suppose to support anything
       def respond_to_missing?(*_args)
         true
-      end
-
-      # Yields the code pretending it is in a transaction
-      # Supports our aborting transaction flow
-      def transaction
-        result = nil
-        commit = false
-
-        catch(:abort) do
-          result = yield
-          commit = true
-        end
-
-        commit || raise(WaterDrop::Errors::AbortTransaction)
-
-        result
-      rescue StandardError => e
-        return if e.is_a?(WaterDrop::Errors::AbortTransaction)
-
-        raise
       end
 
       # @param _args [Object] anything really, this dummy is suppose to support anything